+++ changelog.md
@@ -1,6 +1,12 @@
 jQuery UI/Mobile Sliderbutton Changelog
 =======================================
 
+Version 1.2.0 - (Released: 2012-10-12)
+-------------
+* Adds grunt.js build system.
+* Changes namespace of sliderbutton widget from 'ui'/'mobile' to 'ju'
+* Replaces the separate jQuery simulate ext files with the complete version (all combined)
+
 Version 1.1 - (Released: 2012-08-22)
 -----------
 * Unified the event triggering of mobile and UI version: Both version now behave exactly the same concerning the time and number of events triggered.